feat: 행사관리자가 QR 티켓 강제 재발급하는 로직 구현
