+++ src/main/java/com/fairing/fairplay/qr/controller/QrTicketController.java
@@ -1,6 +1,8 @@
 package com.fairing.fairplay.qr.controller;
 
 
+import com.fairing.fairplay.qr.dto.QrTicketReissueRequestDto;
+import com.fairing.fairplay.qr.dto.QrTicketReissueResponseDto;
 import com.fairing.fairplay.qr.dto.QrTicketRequestDto;
 import com.fairing.fairplay.qr.dto.QrTicketResponseDto;
 import com.fairing.fairplay.qr.dto.QrTicketUpdateRequestDto;
@@ -34,11 +36,36 @@ public class QrTicketController {
     return ResponseEntity.ok(qrTicketService.issueGuest(token));
   }
 
-  // QR 티켓 재발급
+  // 1. QR 티켓 재발급 - QR 티켓 조회 시 이미지코드와 수동 코드만 재발급(새로고침버튼)
   @PostMapping("/reissue")
   public ResponseEntity<QrTicketUpdateResponseDto> reissueQrTicket(
       @RequestBody QrTicketUpdateRequestDto dto) {
     return ResponseEntity.ok(qrTicketService.reissueQrTicket(dto));
+  }
+
+  // 2. QR 티켓 재발급 - 회원이 QR 티켓 분실 시 회원이 티켓 자체 재발급 요청
+  /*
+  * 마이 페이지 접근 가능한 경우
+  * 1. "관리자 문의"
+  * 2. 관리자 승인 후 발급
+  * 3. 이메일로 발급 완료 메일 전송
+  * */
+//  @PostMapping("/admin/reissue")
+//  public ResponseEntity<QrTicketReissueResponseDto> requestReissueQrTicket(@RequestBody QrTicketRequestDto dto) {
+//    return ResponseEntity.ok(qrTicketService.reissueAdminQrTicket());
+//  }
+
+
+  // 3. QR 티켓 강제 재발급 - 비회원/회원 일부 QR 티켓 분실 시 이메일 재전송
+  /*
+   * 마이 페이지도 접근 안될 경우
+   * 1. "관리자 문의"
+   * 2. 관리자 승인 후 발급
+   * 3. 이메일로 발급 완료 메일 전송
+   * */
+  @PostMapping("/admin/reissue/send-email")
+  public ResponseEntity<QrTicketReissueResponseDto> reissueAdminQrTicket(@RequestBody QrTicketReissueRequestDto dto){
+    return ResponseEntity.ok(qrTicketService.reissueAdminQrTicket(dto));
   }
 
   // QR 티켓 만료
